# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Train Schedule and Reservation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trains:
    get:
      operationId: listTrains
      tags: [trains]
      summary: List all trains
      responses:
        "200":
          description: List of trains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Train"
    post:
      operationId: createTrain
      tags: [trains]
      summary: Create a train
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainBody"
      responses:
        "201":
          description: Train created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Train"
  /trains/{trainId}:
    put:
      operationId: updateTrain
      tags: [trains]
      summary: Update a train
      parameters:
        - name: trainId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainBody"
      responses:
        "200":
          description: Train updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Train"
    delete:
      operationId: deleteTrain
      tags: [trains]
      summary: Delete a train
      parameters:
        - name: trainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List all schedules
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBody"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{scheduleId}:
    put:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScheduleBody"
      responses:
        "200":
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: scheduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Train:
      type: object
      properties:
        trainId:
          type: integer
        trainName:
          type: string
        seatCapacity:
          type: integer
      required: [trainId, trainName, seatCapacity]
    CreateTrainBody:
      type: object
      properties:
        trainName:
          type: string
        seatCapacity:
          type: integer
      required: [trainName, seatCapacity]
    Schedule:
      type: object
      properties:
        scheduleId:
          type: integer
        trainId:
          type: integer
        trainName:
          type: string
        origin:
          type: string
        destination:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        date:
          type: string
        availableSeats:
          type: integer
        ticketPrice:
          type: number
      required: [scheduleId, trainId, trainName, origin, destination, departureTime, arrivalTime, date, availableSeats, ticketPrice]
    CreateScheduleBody:
      type: object
      properties:
        trainId:
          type: integer
        origin:
          type: string
        destination:
          type: string
        departureTime:
          type: string
        arrivalTime:
          type: string
        date:
          type: string
        ticketPrice:
          type: number
      required: [trainId, origin, destination, departureTime, arrivalTime, date, ticketPrice]
...